\name{episplineDensity-internal}
\title{Internal Epispline Functions}
\description{
    Internal Epispline Functions
}
\details{
   These are for internal use.
}
\alias{n10}
\alias{e10}
\alias{coeff}
\alias{llfcn}
\alias{llfcn.vec}
\alias{postproc}
\alias{epispline}
\alias{grad_expepispline}
\alias{nonlinear.constraints}
\alias{integrate.to.one}
\alias{generateLinCon}
\alias{obj.nloptr}
\alias{solveML}
\keyword{internal}
