\name{print.episplineDensity}
\alias{print.episplineDensity}
\title{
Print method for episplineDensity objects.
}
\description{
This prints the \code{status} item and nothing else.
}
\usage{
\method{print}{episplineDensity} (x, \ldots)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{
Output from a call to \code{\link{expepi}}.
}
  \item{\dots}{
Other arguements, currently ignored.
}
}
\details{
The current intent of this is to keep this whole object from
printing to the screen.
}
\value{
None
}
\author{
Sam Buttrey
}
\seealso{
\code{\link{expepi}}
}
